Add files via upload
Solution: 
In the below python program takes a list of transactions as inputs, where each transaction is represented by a tuple that contains the transaction type buy or sell, the number of shares and the price share. It then calculates the total capital gain or loss for the sequence of transactions using the FIFO protocol to identify the shares. It then prints the total capital gain. The function utilizes a deque to implement FIFO behavior for selling shares.