Add files via upload
Explain how to use an AVL tree or a red-black tree to sort n comparable elements in O(n logn) time in the worst case.
To sort n comparable elements in O(n log n) time in the worst case using an AVL tree or a red-black tree, we can leverage the self-balancing properties of these data structures. The approach is similar to priority-queue sorting, where elements are inserted into the tree one by one, and then an in-order traversal is performed to retrieve the elements in sorted order.